game_libs_func_00009D24: drop s8-param sign-extend+spill (81.0->89.7%)
37-insn init+unrolled-copy. The `s8 arg3` param made IDO sign-extend
(sll/sra 24) and spill it (`sw a3,12(sp)`) — 3 leading insns absent from
the target, which uses arg3 directly. Declaring arg3 `s32` removes all
three (the `*(s8*)` cast store still emits `sb`). fuzzy 81.0 -> 89.7%.
Residual is the main unrolled loop's induction form: target strength-
reduces to a pointer-end compare (`t0 != a2+arg3*4`) vs my index `i!=arg3`;
a full pointer-walk C rewrite REGRESSED to 58% (lost the remainder-loop
unroll), so kept the index form. Still NON_MATCHING.

+++ src/game_libs/game_libs_tail.c
@@ -491,7 +491,7 @@ int game_libs_func_00009D00(int *a0, int a1) {
 #ifndef FW
 #define FW(p, o) (*(int *)((char *)(p) + (o)))
 #endif
-void game_libs_func_00009D24(char *arg0, s32 arg1, s32 arg2, s8 arg3) {
+void game_libs_func_00009D24(char *arg0, s32 arg1, s32 arg2, s32 arg3) {
     s32 i;
 
     FW(arg0, 0x1C) = arg1;
